% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{A function that estimats B OLS in matrix form}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A numerical vector of size (nx1).}

\item{X}{A numerical matrix of size (nxk).}
}
\value{
A numerical vector of size (kx1).
}
\description{
A function that estimats B OLS in matrix form
}
