% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plottheme.r
\name{my_plottheme}
\alias{my_plottheme}
\title{A function that customizes my plots using my favourite gaury theme}
\usage{
my_plottheme(plot_in)
}
\arguments{
\item{plot_in}{The plot to customize.}
}
\value{
A customized plot.
}
\description{
A function that customizes my plots using my favourite gaury theme
}
